requirements.txt
pyyaml
python-box
beautifulsoup4
requests
pandas
